FHIR: Robust test setup via test spring context.
