+++ src/main/java/org/snomed/snowstorm/config/elasticsearch/IndexConfig.java
@@ -1,10 +1,15 @@
 package org.snomed.snowstorm.config.elasticsearch;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 public class IndexConfig {
 
 	private final String indexNamePrefix;
 	private final short indexShards;
 	private final short indexReplicas;
+
+	private final Logger logger = LoggerFactory.getLogger(getClass());
 
 	public IndexConfig(String indexNamePrefix, short indexShards, short indexReplicas) {
 		if (indexShards < 1) {
@@ -13,6 +18,7 @@ public class IndexConfig {
 		this.indexNamePrefix = indexNamePrefix;
 		this.indexShards = indexShards;
 		this.indexReplicas = indexReplicas;
+		logger.info("Index name prefix: '{}'", indexNamePrefix);
 	}
 
 	public String getIndexNamePrefix() {
